requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.11
colorlog==6.6.0
frozenlist==1.3.0
hikari==2.0.0.dev106
hikari-lightbulb==2.2.0
idna==3.3
multidict==6.0.2
Pillow==9.0.1
yarl==1.7.2
